bouncebuf: Allow printf-simple to debug messages.
printf-simple does not have size_t-support (format string '%zu')
with the effect, that the debug output will just contain a 'u'
instead of the provided number.

This patch converts the size_t value to unsigned long and use
the format string %lu to print it. With that the debug messages
can be printed with printf-simple.

+++ common/bouncebuf.c
@@ -22,7 +22,7 @@ static int addr_aligned(struct bounce_buffer *state)
 
 	/* Check if length is aligned */
 	if (state->len != state->len_aligned) {
-		debug("Unaligned buffer length %zu\n", state->len);
+		debug("Unaligned buffer length %lu\n", (unsigned long)state->len);
 		return 0;
 	}
